LayoutTests/http/wpt/wk-web-animations/timing-model/animation-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that the Animation currentTime property is set based on the timeline currentTime and the animation startTime</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
    const animation = new Animation(document.timeline);
    assert_equals(animation.currentTime, null);
}, "An animation's currentTime is null when the timeline has no currentTime set.");

test(t => {
    const animation = new Animation(document.timeline);
    animation.startTime = 2;
    internals.setTimelineCurrentTime(document.timeline, 0);
    assert_equals(animation.currentTime, -2);
    internals.setTimelineCurrentTime(document.timeline, 3);
    assert_equals(animation.currentTime, 1);
    animation.startTime = 1;
    assert_equals(animation.currentTime, 2);
}, "An animation accounts for its start time and document current time when computing its current time.");

test(t => {
    const animation = new Animation(document.timeline);
    animation.startTime = 1;
    internals.setTimelineCurrentTime(document.timeline, 3);
    assert_equals(animation.currentTime, 2);

    animation.currentTime = 1;
    assert_equals(animation.startTime, 2);
}, "Setting the current time should adjust the start time.");

test(t => {
    const animation = new Animation(document.timeline);
    assert_throws(new TypeError, () => animation.currentTime = null);
}, "Setting the current time to null should throw a TypeError.");

</script>
</body>
